fix: Extracting GAPIC API calls into a new module
